+++ CHANGELOG.md
@@ -11,7 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `forecast_5_day` method to enable 5 day forecast requests based on either
   location name or zip/postal code.
-  
+
+- `forecast_hourly` method to enable hourly forecast requests based on either
+  location name or zip/postal code.
+
 ### Deprecated
 
 - `get_location_weather` method will be deprecated in version 1.0.0.